Update patch set 4
Patch Set 4:

My personal plea:
Please don't put the actual firmware binaries in the EC repository,
this will end up making the git repository large and slow (and we will realize that we have submitted too many revisions of those binaries, it will be too late as we don't want to nuke the repo).
Overall you can definitely submit an updated url/text config for each config you want to release to point to the current stable version, but putting the actual binary somewhere else like the Google storage buckets we are using for everything else is usually the way to go.

Patch-set: 4
CC: Gerrit User 1000016 <1000016@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>