Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Maybe it's more user-friendly to just print in degrees C?

We could, but then we'd have to make the tests (see linked CL and comment on the bug) a bit smarter to handle the fact that having no unit means K...

Also, the data is also available from /sys/class/thermal/thermal_zone*/temp, in milli-celsius, so that should be good enough.

My main point here is to make sure people do not get confused when reading the data, and assume that 300 is 30.0 C (this happened to Roy).

Patch-set: 1
Reviewer: Gerrit User 1129199 <1129199@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>