Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Does this need to be disabled for release? I suppose someone would have to have physical access, so this is fine?

I don't think so: The code indicates that the original author intends it to be always enabled. All this does is to make it actually work for the STM32F0x MCUs. If there are concerns we could wrap it in a CONFIG_ parameter, but for the other cores where it supposedly works in the current state, there is no such thing.

Patch-set: 1
Reviewer: Gerrit User 1177024 <1177024@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>