chip/stm32/usb_hid_keyboard: add dynamic backlight detection support
In order to make a single hammer image support both base with and
without keyboard backlight. We need a way to dynamically determine if
backlight is present, and send the correct HID descriptors. This is done
through declaring two HID descriptors with equal length, with one
including HID output descriptor information and the other one does not.

BUG=b:67722756
TEST=manually

Change-Id: Ie3eac9b3d4cd749308ccfb96a7db469701f9793b