Update patch set 1
Patch Set 1:

I'm not putting a negative score because I'm not a reviewer here.
But USB_PD_LOGGING is the only way of debugging PD issues in the field, so if you need so dramatically more flash space to do this, this board support is probably dead for master and future products.

As a consequence, now I'm very curious what this new feature brings to the table since it's written nowhere in b/69329196, so it worth crippling other boards.

And yes, it's costly to do 64-bit bitmaps (even more using the ARMv6-M ISA which has poor support) that's why we were not doing it anywhere else.

Patch-set: 1
CC: Gerrit User 1000016 <1000016@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>