Update patch set 1
Patch Set 1:

> I'm not putting a negative score because I'm not a reviewer here.
But USB_PD_LOGGING is the only way of debugging PD issues in the field, so if you need so dramatically more flash space to do this, this board support is probably dead for master and future products.
>
> As a consequence, now I'm very curious what this new feature brings to the table since it's written nowhere in b/69329196, so it worth crippling other boards.
>
> And yes, it's costly to do 64-bit bitmaps (even more using the ARMv6-M ISA which has poor support) that's why we were not doing it anywhere else.

Yes, Shawn actually talked about getting rid of just "gru" since it might not even be booting off ToT at the moment. But, since it could be done as a future cleanup, we did not handle it right away.

Reason for enabling 64-bit hostevents is because we are already out of available bits in the current 32-bit hostevents and any future need for new bits would require us recycling the current events and thus breaking compatibility with older versions. Additionally, we are currently working on a new command to enable programming of hostevents/wakemasks. So, we took this as an opportunity to increase the event size so that all the old hostevent commands will continue operating on lower 32-bits to maintain compatibility with older OS and BIOS versions. On the other hand, the new command can support 64-bit host events and allow newer BIOS and kernel versions to take advantage of the new available event bits.

Patch-set: 1
Reviewer: Gerrit User 1003000 <1003000@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Trybot-Ready=0