charge_manager: Add batt.is_present to check for exiting safe_mode
When recovering from ship mode a battery may have the D-FET disabled
until the charger can provide it with some charging current. In this
case if safe_mode is exited prior to battery being able to provide
current, then it can result in calling board_set_active_charge_port()
with CHARGE_PORT_NONE as the parameter. For the case of the Rohm
charger this results in VSYS getting shutdown.

Adding a check for batt.is_present to the exit safe_mode criteria
ensures that the battery is able to provide power for the system
before the charger exits safe mode.

BUG=chromium:777596
BRANCH=None
TEST=Tested on Coral (Nasher with LG battery). Without this change,
observed that board_set_active_charge_port() is consistently called
with a -1 argument resulting in the system powering down. Adding
batt.is_present results in consistent power on operation when
recovering from battery ship mode condition.

Change-Id: Ibe2659017bcbaabd77cdf455febf68bab80879ee
Signed-off-by: Scott Collyer <scollyer@google.com>