Update patch set 16
Patch Set 16:

Mario, 

Running 'accelinfo on' on a nucleos f0, I notice the units are not right:
[1049.700254 event 0x80000000 LSM6DSM ACC=-57  , 110  , 1016 ...

The values reported here and to the host are not in G (anymore).
The range [2g] defines that 1<<15 would be 2g, and -(1<<15) will be -2g.
So instead of ~1000 on the Z axis, I would have expected ~1<<14, ~16000.

I already fixed the barometer, I am guessing all other sensors are affected.

Patch-set: 16
Reviewer: Gerrit User 1104704 <1104704@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>