Update patch set 2
Patch Set 2:

yeah, this is the part I was not fully clear about. Why would we care about the password and stuff, if we want to lock up the device?

Should it be sufficient to check that it is in RMA mode and lock it if so? This command should be usable only if the device was unlocked for RMA. Or is the issue that there is no way to know how the device was unlocked?

Patch-set: 2
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>