driver: lsm6dsm/lsm6dsl: Fix acc. units
The range [2g] defines that acc data 1<<15 would be 2g,
and -(1<<15) will be -2g. Actually accelerometer
report units in mg.
Fixed gain in order to have for 1g a value of ~16384.

BUG=none
BRANCH=master
TEST=Tested on discovery BOARD with LSM6DSM sensor connected
to I2C master interface of target board.
Using accelrange motion sense console command is possible to
test with different full scale settings. With 2g full scale
running "accelinfo on 1000" command we can read the following
sensor data:

[29521.969665 event 0x00000005 LSM6DSM ACC=-264 , -824 , 16880]

Selecting a different scale to 4g also we can read the same value:

[29680.655558 event 0x00000005 LSM6DSM ACC=-234 , -780 , 16788]

and so on with 8g and 16g full scale.

Change-Id: I9f7b08374df6f1da14dcc6917751e8e94f0b9d5c
Signed-off-by: Mario Tesi <mario.tesi@st.com>