g:sps: do not stoll reset when CS is asserted
Stalling reset during when CS is asserted is useful to start with, it
was added before out of abundance of caution, but come to think of it,
should the reset happen asynchronously driven be the EC, the AP would
be reset too. And when AP is reset on its own accord, it would not be
transmitting anything on the SPI interface.

On top of that it turns out that in some cases reset on ARM platforms
is accompanied by the CS line driven low, which causes infinite loop
if Cr50 is waiting for CS to deassert before proceeding.

BRANCH=cr50
BUG=b:69236269
TEST=verified that RMA reset operates properly on both ARM and x86
     platforms.

Change-Id: I5e99bb6472c10c88ddfb48d12cddcbb8b03b7953
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>