power: introducing pwr_avg console command
Keep a running average of power consumption, as seen by the ec
talking to the battery, over a 1 minute span. Samples get collected
every second.
This change allows us to have better power tracking while minimizing
the power impact on the EC, because
- the pwr_avg command only needs to be called once every minute, and is
short, thus less expensive to parse on ECs without a UART buffer
- the recurring update is fairly inexpensive, since it only asks the
battery for current and voltage.

change to chell is due to size constraint
changes to samus/wheatley are due to this functionality
not being necessary on those boards.

usage:
> pwr_avg
mw = -441

BUG=
TEST=make buildall -j

Change-Id: Id1a3479d277aedf90dfa965afb4ee9136654b1cf
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>