WIP, power: introducing pwr_avg console command
pwr_avg provides an average voltage, amperage, and wattage over the last
1 minute. It's up to the battery drivers to impliment this
functionality.
This change allows us to have better power tracking while minimizing
the power impact on the EC, because
- the pwr_avg command only needs to be called once every minute, and is
short, thus less expensive to parse on ECs without a UART buffer
- the work done to keep the avg is partially done by the batteries
already and it's just a question of retrieving it.

WIP until:
- I find space in chell ;)
- waiting to hear back about maxim's avg calculation algorithm
- someone verifies pwr_avg on Dru (the only non-smart driver)
- I run voltage tests to see if we really need to sample volt every 10s,
or if alternatives, like sampling every Xs (where X >> 10), or just
taking the average between two readings.

changes to samus/wheatley are due to this functionality
not being necessary on those boards.

usage:
> pwr_avg
mw = -441

BUG=chromium:752320
BRANCH=None
TEST=make buildall -j

Change-Id: Id1a3479d277aedf90dfa965afb4ee9136654b1cf
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>