grunt: Add accel, gyro, and motion_sense task.
This adds support for the KX022 and BMI160 sensors.

It also enables the motion_sense task so that the keyboard can be
disabled in tablet mode.

This change doesn't deal with interrupt handling for the BMI160 part; it
seems big enough as it is and that can be dealt with separately.

BUG=b:69379570,b:69379621
BRANCH=None
TEST=Build

Change-Id: I72cc8f86ddd8d466b4208ff583e93fd51da63c49
Signed-off-by: Justin TerAvest <teravest@chromium.org>