panic: Set EC_HOST_EVENT_PANIC on chipset reset
Add a hook for CHIPSET_RESET to allow the EC to indicate if there is
any new panic info present. This helps coreboot to log EC panic info
in eventlog.

Also, update the hook priority for CHIPSET_RESET and HOOK_INIT to
HOOK_PRIO_LAST to allow the EC to first log any software panic before
it is checked.

BUG=b:65732924,b:69334392
BRANCH=None
TEST=Verified following:
1. Force panic_set_reason in EC on CHIPSET_RESET
2. reboot on AP console
3. mosys eventlog list shows "EC Event | Panic Reset in previous boot"

Change-Id: I314539a724a3f23cd54ebfb8114ff64ee6ff2ce7
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 9fa6de15cf57e4beaa5726ebbe0e4ac6af004a79
Original-Change-Id: I77b49cd0b3bf05b10efc708e3d81af9ed0e3aa49
Original-Signed-off-by: Furquan Shaikh <furquan@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/797911
Original-Reviewed-by: Shawn N <shawnn@chromium.org>