coral: Add request nil override for Nasher BYD and LG batteries
To avoid inadvertenly shutting off the charger, for these 2 batteries
if the battery SOC is 0% and the battery requests 0 for voltage and
current, then override this request with precharge current and max
voltage to ensure that the battery wakes up.

BUG=b:67018220
BRANCH=coral
TEST=Connected external passive load and drained BYD battery to point
where the D-FET gets disabled (~8.6v). Then connect external charger
and verify that battery does requrest 0V/0A and that the override is
active and the battery wakes up correctly.

Change-Id: Iff9ceffa010ac3e14107220f2a0c6eb475c1e46d
Signed-off-by: Scott Collyer <scollyer@google.com>