PPC: Add support for shared IRQ.
This commit adds support for a PPC that shares its interrupt line with
the TCPC.  This can be enabled by the following CONFIG_* option:

    CONFIG_USBC_PPC_SHARED_IRQ

If a board uses this config, the interrupt should be level sensitive
instead of edge triggered.  The TCPC will be serviced first, followed by
the PPC.

BUG=b:69139844
BRANCH=None
TEST=Flash meowth; Verify that PD negotiaton still works, verify that
overcurrent detection still works properly.

Change-Id: I354892574746e80c050a57b401a2bd28df99e7a0
Signed-off-by: Aseda Aboagye <aaboagye@google.com>