cr50: preserve RMA mode in NVMEM
Eve Cr50 branch is lacking CCD implementation, so a separate means of
preventing WP state while in RMA mode is needed.

Let's add an NVMEM variable to serve as a flag indicating that WP
should not be enabled. The variable is set when RMA mode is entered
and cleared when RMA mode is exited.

BRANCH=none
BUG=none
TEST=verified that WP is not active even after removing and
     re-attaching the battery while the device is in RMA mode.

Change-Id: Ib75bd7f29d48bea01c4b962c22fb2a14706efb96
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>