Update patch set 3
Patch Set 3:

Thanks Aseda, Shawn for your comments. 
Sorry for the delay in response.
This is a valid point. The SMC_SHUTDOWN GPIO from EC to NC7SZ(flip-flop) determines the fate of A_RAIL_EN. Which indicates that power to EC will also go off, in that case Nuvoton WTC (wake time counter) register cannot be programmed and it will not wake up after specified seconds.
The only reason this test case worked was because AC adapter was plugged in and in glkrvp SMC_SHUTDOWN pin has no affect as the logic is ORed with AC_ADPTR_OK. 
When tested with only battery connected this fails, as there is no Vcc1 for Nuvoton to configure WTC registers. 
The solution to this would be to remove the SMC_SHUTDOWN from the hibernate function along with the while loop.
Let me know your thoughts.

Patch-set: 3