Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)

It can be, but the g78x driver would need a some work. The registers are different and I think the driver assumes 2 temp sensors on the g78x (one internal, one external) which the FT4 doesn't have. It's also not a g78x so I wasn't sure if that was the right place for it or not. I was going to finish this up today though, so let me know if you think they should be merged.

Patch-set: 3
Reviewer: Gerrit User 1242164 <1242164@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>