Update patch set 3
Patch Set 3:

Oh my mistake, I was confusing the sensor internal to the G781 with the sensor internal to the FT4. And it sounds like the differences between the G78x and the FT4 are enough to keep the drivers separate.

Patch-set: 3
Reviewer: Gerrit User 1212794 <1212794@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>