cr50: Don't touch resets for USB-SPI bridge.
Cr50 should not automatically touch the device resets when enabling the
USB-SPI bridge.  Otherwise, this could interefere with devices that
might have internal SPI flash and need to have their resets deasserted
in order to access the internal SPI flash.

This commits simply removes any deasserting or asserting of the resets
when enabling or disable the USB-SPI bridge.  The user or external
scripts should control the resets as necessary using servo or the cr50
console.

BUG=b:71548795,b:71557464
BRANCH=None
TEST=Flash meowth cr50.  Verify that I can flash the EC using a
servo_v4.
TEST=Repeat above test with a servo_micro.
TEST=Repeat above test with a SuzyQable.

Change-Id: I114c34df43cf1e8ba622e75c3e6ecf517afc40a4
Signed-off-by: Aseda Aboagye <aaboagye@google.com>