Update patch set 1
Patch Set 1:

I am not sure this is such a good idea: we sure must keep x86 chips in reset when programming AP flash, at least so far we never had an ARM AP with a built-in SPI flash, so to be consistent with the current use across the board (like being able to just use flashrom) it is better left alone.

Not so sure about the EC, but then again, it would be great to be able to program just by running flashrom, without touching anything else. That is a different programming mode for this EC could be a better option, Let's see what others think.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>